refactor: rename gpt_oss eagle example dir to peagle
Rename examples/models/gpt_oss/eagle/ to peagle/ to reflect the
parallel-EAGLE (P-EAGLE) drafter it implements, updating the eagle.*
package imports and path references accordingly. Algorithm-name
mentions (eagle3, get_eagle_config) are left unchanged.
