+++ examples/models/gpt_oss/peagle/drafter_cpu.py
