+++ examples/models/gpt_oss/peagle/drafter_model.py
