+++ examples/models/gpt_oss/peagle/kernels/drafter_layer.py
