+++ examples/models/gpt_oss/peagle/kernels/rope.py
