+++ examples/models/gpt_oss/peagle/kernels/softmax.py
