+++ examples/models/gpt_oss/peagle/tensor_preparation.py
